% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devOutliers.R
\name{eventVSEnvironmental}
\alias{eventVSEnvironmental}
\title{Get Environnemental Data and events}
\usage{
eventVSEnvironmental(scientificObjectURI, startDate = "", endDate = "",
  showPoint = FALSE, pageSize = 1000)
}
\arguments{
\item{scientificObjectURI}{the object which we will retreive the event}

\item{startDate}{the begining of the period}

\item{endDate}{the end of the period}

\item{showPoint}{show points
#@param variableList from \code{\link{variableList}}
#@param token a token from \code{\link{getToken}} function}

\item{pageSize}{the size of data retreived by default 1000 measures}
}
\value{
plot
}
\description{
Get Environnemental Data and events
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
 aToken <- getToken("guest@opensilex.org","guest")
 token <- aToken$data
 startDate ="2017-05-22"
 endDate ="2017-06-30"
 pageSize=20000
 scientificObjectURI="http://www.opensilex.org/demo/2018/o18000175"
 showPoint = FALSE
 envEvents::eventVSEnvironmental(scientificObjectURI, startDate, endDate,showPoint,pageSize)
}
}
